+++ android/app/src/debug/AndroidManifest.xml
@@ -4,4 +4,14 @@
          to allow setting breakpoints, to provide hot reload, etc.
     -->
     <uses-permission android:name="android.permission.INTERNET"/>
+
+    <service
+    android:name=".MyFirebaseMessagingService"
+    android:exported="false"
+    android:directBootAware="true">
+    <intent-filter>
+        <action android:name="com.google.firebase.MESSAGING_EVENT" />
+    </intent-filter>
+</service>
+
 </manifest>
