+++ lib/firebase_options.dart
@@ -3,7 +3,7 @@
 import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
 import 'package:flutter/foundation.dart'
     show defaultTargetPlatform, kIsWeb, TargetPlatform;
-import 'firebase_options.dart';
+// import 'firebase_options.dart';
 
 /// Default [FirebaseOptions] for use with your Firebase apps.
