+++ pubspec.yaml
@@ -31,13 +31,17 @@ dependencies:
   flutter:
     sdk: flutter
   shared_preferences: ^2.2.2
-
+  table_calendar: ^3.0.6
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   firebase_core: ^2.25.3
   firebase_auth: ^4.17.3
+  firebase_database: ^10.4.4
+  firebase_messaging: ^14.7.14
+  flutter_local_notifications: ^16.3.2
+
 
 dev_dependencies:
   flutter_test:
